--- pitch .py
@@ -1,3 +1,0 @@
-from sistema import PitchSecurity as PS
-
-PS()
