+++ pitch.py
@@ -1,0 +1,3 @@
+from sistema import PitchSecurity as PS
+
+PS()
